# GGRC-5521 improve evidence reindex speed