--- a/src/ggrc/models/evidence.py
+++ b/src/ggrc/models/evidence.py
@@ -3,7 +3,10 @@
 
 """Module containing Evidence model."""
 
+from sqlalchemy import and_
 from sqlalchemy import orm
+from sqlalchemy import or_
+from sqlalchemy.ext.declarative import declared_attr
 
 from ggrc import db
 from ggrc import login
@@ -19,7 +22,7 @@ from ggrc.models.mixins import before_flush_handleable as bfh
 from ggrc.models.mixins import base
 from ggrc.models.mixins.statusable import Statusable
 from ggrc.models.mixins.with_auto_deprecation import WithAutoDeprecation
-from ggrc.models.relationship import Relatable
+from ggrc.models.relationship import Relatable, Relationship
 from ggrc.utils import referenced_objects
 
 from ggrc.services import signals
@@ -28,8 +31,8 @@ from ggrc.services import signals
 class Evidence(Roleable, Relatable, mixins.Titled,
                bfh.BeforeFlushHandleable, Statusable,
                mixins.WithLastDeprecatedDate, comment.Commentable,
-               WithAutoDeprecation, mixin.Indexed, base.ContextRBAC,
-               mixins.Slugged, db.Model):
+               WithAutoDeprecation, base.ContextRBAC,
+               mixins.Slugged, mixin.Indexed, db.Model):
   """Evidence (Audit-scope URLs, FILE's) model."""
   __tablename__ = "evidence"
 
@@ -124,20 +127,31 @@ class Evidence(Roleable, Relatable, mixins.Titled,
     return kind
 
   @classmethod
-  def indexed_query(cls):
-    return super(Evidence, cls).indexed_query().options(
-        orm.Load(cls).undefer_group(
-            "Evidence_complete",
-        ),
+  def _populate_query(cls, query):
+    return query.options(
+      orm.subqueryload(cls._related_assessment),
+      orm.subqueryload(cls._related_audit).load_only("archived"),
+      orm.Load(cls).undefer_group(
+        "Evidence_complete",
+      ),
     )
+
+  @classmethod
+  def indexed_query(cls):
+    return cls._populate_query(super(Evidence, cls).indexed_query())
+
+  @classmethod
+  def eager_query(cls):
+    return cls._populate_query(super(Evidence, cls).eager_query())
 
   @simple_property
   def archived(self):
-    """Returns a boolean whether parent is archived or not."""
-    parent_candidates = self.related_objects(_types=Evidence._allowed_parents)
-    if parent_candidates:
-      parent = parent_candidates.pop()
-      return parent.archived
+    """Evidence archived if related Assessment/Audit is archived"""
+    # pylint: disable=unsubscriptable-object
+    if self._related_assessment:
+      return self._related_assessment.audit.archived
+    elif self._related_audit:
+      return self._related_audit.archived
     return False
 
   def log_json(self):
@@ -268,3 +282,81 @@ class Evidence(Roleable, Relatable, mixins.Titled,
   def handle_before_flush(self):
     """Handler that called  before SQLAlchemy flush event"""
     self._map_parent()
+
+  @declared_attr
+  def _related_audit(cls):  # pylint: disable=no-self-argument
+    """Audits mapped to Evidence"""
+    def primary_join_function():
+      return or_(
+          and_(
+              Relationship.source_id == cls.id,
+              Relationship.source_type == cls.__name__,
+              Relationship.destination_type == "Audit"
+          ),
+          and_(
+              Relationship.destination_id == cls.id,
+              Relationship.destination_type == cls.__name__,
+              Relationship.source_type == "Audit"
+          )
+      )
+
+    def secondary_join_function():
+      from ggrc.models import all_models
+      return or_(
+          and_(
+              all_models.Audit.id == Relationship.destination_id,
+              Relationship.destination_type == "Audit",
+          ),
+          and_(
+              all_models.Audit.id == Relationship.source_id,
+              Relationship.source_type == "Audit",
+          )
+      )
+
+    return db.relationship(
+        "Audit",
+        primaryjoin=primary_join_function,
+        secondary=Relationship.__table__,
+        secondaryjoin=secondary_join_function,
+        viewonly=True,
+        uselist=False
+    )
+
+  @declared_attr
+  def _related_assessment(cls):  # pylint: disable=no-self-argument
+    """Assessments mapped to Evidence"""
+    def primary_join_function():
+      return or_(
+          and_(
+              Relationship.source_id == cls.id,
+              Relationship.source_type == cls.__name__,
+              Relationship.destination_type == "Assessment"
+          ),
+          and_(
+              Relationship.destination_id == cls.id,
+              Relationship.destination_type == cls.__name__,
+              Relationship.source_type == "Assessment"
+          )
+      )
+
+    def secondary_join_function():
+      from ggrc.models import all_models
+      return or_(
+          and_(
+              all_models.Assessment.id == Relationship.destination_id,
+              Relationship.destination_type == "Assessment",
+          ),
+          and_(
+              all_models.Assessment.id == Relationship.source_id,
+              Relationship.source_type == "Assessment",
+          )
+      )
+
+    return db.relationship(
+        "Assessment",
+        primaryjoin=primary_join_function,
+        secondary=Relationship.__table__,
+        secondaryjoin=secondary_join_function,
+        viewonly=True,
+        uselist=False
+    )