--- a/test/integration/ggrc/fulltext/test_total_reindex.py
+++ b/test/integration/ggrc/fulltext/test_total_reindex.py
@@ -36,7 +36,7 @@ class TestTotalReindex(TestCase):
       'CycleTaskEntry': 4,
       'CycleTaskGroup': 5,
       'CycleTaskGroupObjectTask': 5,
-      'Evidence': 25,   # TODO improve
+      'Evidence': 7,
       'Document': 5,
       'Issue': 8,
       'Market': 8,
@@ -78,6 +78,7 @@ class TestTotalReindex(TestCase):
       ggrc_factories.ProgramFactory,
       ggrc_factories.RegulationFactory,
       ggrc_factories.DocumentFactory,
+      ggrc_factories.EvidenceFactory,
       ggrc_factories.MetricFactory,
       ggrc_factories.ProductFactory,
       ggrc_factories.ProductGroupFactory,