# Issue description
Currently reindex of 6k evidence takes about 5 min at local env. (25 queries for reindex 10 evidence)

# Steps to test the changes

Compare reindex speed at DB with 6k+ evidence

# Solution description

Updated indexed query to load related assessments and audits

# Performance comparation
Local env (db tempfs disabled)

**Before (dev)**
ggrc.views Evidence: 6160 / 6160
ggrc.performance **112.15s** (**67.99s** CPU and 0.00s for 1  db queries) 'POST /_background_tasks/reindex'

**After**
ggrc.views Evidence: 6160 / 6160
ggrc.performance **73.67s** (**23.24s** CPU and 0.00s for 1  db queries) 'POST /_background_tasks/reindex'


# Sanity checklist

- [x] I have clicked through the app to make sure my changes work and not break the app.
- [x] I have applied the correct milestone and labels.
- [x] My changes fix the issue described in the description (and do nothing else). 🤞
- [x] My changes are covered by tests.
- [x] My changes follow our [performance guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/performance.rst).
- [x] My changes follow our [js](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/javascript.rst) and/or [python](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/python.rst) guidelines.
- [x] My commits follow our [commit guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/git/how_to_write_a_commit_message.rst).

<!-- If your PR includes a migration include the additional checklist items
# Migration checklist
- [ ] Migration passes all checks from our [PR review guidelines](https://github.com/google/ggrc-core/blob/dev/docs/source/contributing/git/reviewing_pull_requests.rst#reviewing-a-pr-containing-database-migration-scripts)
-->

# PR Review checklist

- [x] The changes fix the issue and don't cause any apparent regressions.
- [x] Labels and milestone are correctly set.
- [x] The solution description matches the changes in the code.
- [ ] There is no apparent way to improve the performance & design of the new code.
- [x] The pull request is opened against the correct base branch.
- [x] Upon merging, the Jira ticket's fixversion is correctly set and the ticket is moved to "QA - In Progress".

<!-- If your code is not finished yet can include a TODO check list
# TODO

- [ ] First item on the TODO
-->