@zidarsk8 2018-09-07
@Smotko let's not worry about those redundant acl columns, as they will be removed in the next step of acl improvements, but those audit and assessment queries would be nice if we could get rid of them.