@olegnatsevsky 2018-09-12
@Smotko Let's merge it as it is. After the next step of acl improvements we can recheck if we can improve evidence reindex.