@NastassiaKiseleva 2018-09-23
@olegnatsevsky please create a ticket here, to proceed with merge